Cambios

+++ sensorHIH-I2C.c
@@ -72,12 +72,12 @@ int main(int argc, char **argv)
 			if((buf[0] & 0xC0) == 0)
 			{	
 				int reading_hum = (buf[0] << 8) + buf[1];
-				double humidity =roundf((reading_hum / 16382.0 * 100.0)*100)/100;
+				double humidity =round((reading_hum / 16382.0 * 100.0)*100)/100;
 				printf("Humidity: %f\n", humidity);
 
 				/* Temperature is located in next two bytes, padded by two trailing bits */
 				int reading_temp = (buf[2] << 6) + (buf[3] >> 2);
-				double temperature = roundf((reading_temp / 16382.0 * 165.0 - 40)*100)/100;
+				double temperature = round((reading_temp / 16382.0 * 165.0 - 40)*100)/100;
 				printf("Temperature: %f\n\n", temperature);
 			}else 
 				printf("Error, el estado es diferente de 0\n");
